+++ lib/stone_bank/operations.ex
@@ -7,6 +7,7 @@ defmodule StoneBank.Operations do
   @transfer "transfer"
   @exchange "exchange"
   @deposit "deposit"
+  @split_payment "split_payment"
 
   alias StoneBank.{Accounts, Accounts.Account}
   alias StoneBank.Repo
@@ -75,9 +76,35 @@ defmodule StoneBank.Operations do
     |> transaction_case("deposit with success!!! from: #{from.id} value: #{value}")
   end
 
-  defp is_negative_amount?(from_amount, value) do
-    {:ok, %Money{currency: _, amount: value_in_money}} = Money.sub(from_amount, value)
-    Decimal.negative?(value_in_money)
+  def split_payment(from, to_accounts_id, value) do
+    value = Money.new(:brl, value)
+
+    case is_negative_amount?(from.amount, value) do
+      true ->
+        {:error, "you can't have negative amount!"}
+
+      false ->
+        perform_operation(to_accounts_id, value, :split_payment)
+        split_payment_operation(from, to_accounts_id, value)
+    end
+  end
+
+  defp split_payment_operation(from, to_accounts_id, value) do
+    to_accounts =
+      to_accounts_id
+      |> Enum.map(&to_string/1)
+      |> Enum.join(", ")
+
+    Ecto.Multi.new()
+    |> Ecto.Multi.update(:account_from, perform_operation(from, value, :sub))
+    |> Ecto.Multi.insert(
+      :transaction,
+      gen_transaction(value, from.id, to_accounts, @split_payment)
+    )
+    |> Repo.transaction()
+    |> transaction_case(
+      "split payment with success!!! from: #{from.id} to: #{to_accounts} value: #{value}"
+    )
   end
 
   defp perform_update(from, t_id, value) do
@@ -115,8 +142,27 @@ defmodule StoneBank.Operations do
     update_account(account, %{amount: value_response_sub})
   end
 
+  defp perform_operation(accounts, value, :split_payment) do
+    Enum.map(accounts, fn ids ->
+      x = Accounts.get!(ids)
+      acc = Money.new(:brl, "0")
+
+      accounts_number = Enum.count(accounts)
+      {:ok, splited_value} = Money.div(Money.add!(acc, value), accounts_number)
+      value_response_sum = Money.add!(x.amount, splited_value)
+
+      update_account(x, %{amount: value_response_sum})
+      |> Repo.update()
+    end)
+  end
+
   def update_account(%Account{} = account, params) do
     Account.changeset(account, params)
+  end
+
+  defp is_negative_amount?(from_amount, value) do
+    {:ok, %Money{currency: _, amount: value_in_money}} = Money.sub(from_amount, value)
+    Decimal.negative?(value_in_money)
   end
 
   defp gen_transaction(value, f_id, t_id, type) do