+++ lib/stone_bank/transactions/transaction.ex
@@ -8,6 +8,7 @@ defmodule StoneBank.Transactions.Transaction do
 
   @primary_key {:id, Ecto.UUID, autogenerate: true}
   @required_params [:value, :account_from, :account_to, :type, :date]
+  @derive {Jason.Encoder, only: [:account_from, :account_to, :date, :type, :value]}
 
   schema "transactions" do
     field :account_from, :string